requirements.txt
paho-mqtt
influxdb
python-twitter
sdnotify
